@@ -1,6 +1,6 @@
 {
-    "name": "horde/imap-client",
-    "description": "IMAP client library",
+    "name": "bytestream/horde-imap-client",
+    "description": "Horde IMAP client library",
     "type": "library",
     "homepage": "https://www.horde.org/libraries/Horde_Imap_Client",
     "license": "LGPL-2.1",
@@ -11,49 +11,33 @@
             "role": "lead"
         }
     ],
-    "version": "2.30.0",
-    "time": "2019-09-29",
-    "repositories": [
-        {
-            "type": "pear",
-            "url": "https://pear.horde.org"
-        }
-    ],
     "require": {
-        "php": "^5.3 || ^7",
-        "pear-pear.horde.org/Horde_Exception": "^2",
-        "pear-pear.horde.org/Horde_Mail": "^2",
-        "pear-pear.horde.org/Horde_Mime": "^2.5.2",
-        "pear-pear.horde.org/Horde_Socket_Client": "^2",
-        "pear-pear.horde.org/Horde_Stream": "^1.4",
-        "pear-pear.horde.org/Horde_Secret": "^2",
-        "pear-pear.horde.org/Horde_Stream_Filter": "^2",
-        "pear-pear.horde.org/Horde_Translation": "^2.2",
-        "pear-pear.horde.org/Horde_Util": "^2",
+        "php": "^7.4 || ^8.0",
+        "bytestream/horde-exception": "^2",
+        "bytestream/horde-mail": "^2",
+        "bytestream/horde-mime": "^2.5.2",
+        "bytestream/horde-socket-client": "^2",
+        "bytestream/horde-stream": "^1.4",
+        "bytestream/horde-secret": "^2",
+        "bytestream/horde-stream-filter": "^2",
+        "bytestream/horde-translation": "^2.2",
+        "bytestream/horde-util": "^2",
         "ext-hash": "*",
         "ext-json": "*"
     },
-    "suggest": {
-        "pear-pear.horde.org/Horde_Cache": "^2",
-        "pear-pear.horde.org/Horde_Compress_Fast": "^1",
-        "pear-pear.horde.org/Horde_Crypt_Blowfish": "^1.1",
-        "pear-pear.horde.org/Horde_Db": "^2.2",
-        "pear-pear.horde.org/Horde_HashTable": "^1",
-        "pear-pear.horde.org/Horde_Mongo": "^1",
-        "pear-pear.horde.org/Horde_Pack": "^1",
-        "pear-pear.horde.org/Horde_Stringprep": "^1",
-        "pear-pear.horde.org/Horde_Support": "^2",
-        "pear-pear.horde.org/Horde_Test": "^2.2.7",
-        "ext-intl": "*",
-        "ext-mbstring": "*"
-    },
-    "replace": {
-        "pear-pear.horde.org/Horde_Imap_Client": "2.*",
-        "pear-horde/Horde_Imap_Client": "2.*"
+    "require-dev": {
+        "phpunit/phpunit": "^9.5",
+        "bytestream/horde-stringprep": "^1.1",
+        "bytestream/horde-crypt-blowfish": "^1.2"
     },
     "autoload": {
         "psr-0": {
             "Horde_Imap_Client": "lib/"
+        }
+    },
+    "autoload-dev": {
+        "psr-0": {
+            "Horde_Imap_Client": "test/"
         }
     }
 }
