@@ -24,10 +24,11 @@
  * @subpackage UnitTests
 
  */
-class Horde_Imap_Client_AuthTest extends PHPUnit_Framework_TestCase
+class Horde_Imap_Client_AuthTest extends \PHPUnit\Framework\TestCase
 {
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
+        parent::setUpBeforeClass();
         require_once __DIR__ . '/Stub/DigestMD5.php';
         require_once __DIR__ . '/Stub/Scram.php';
     }
