@@ -24,7 +24,7 @@
  * @subpackage UnitTests
  */
 abstract class Horde_Imap_Client_Data_Fetch_TestBase
-extends PHPUnit_Framework_TestCase
+extends \PHPUnit\Framework\TestCase
 {
     private $ob;
 
@@ -32,8 +32,9 @@ extends PHPUnit_Framework_TestCase
     protected $ob_class;
     abstract protected function _setUp();
 
-    public function setUp()
+    public function setUp(): void
     {
+        parent::setUp();
         $this->_setUp();
 
         $this->ob = new Horde_Imap_Client_Data_Fetch($this->ob_class);
