@@ -24,7 +24,7 @@
  * @subpackage UnitTests
  */
 class Horde_Imap_Client_Data_SearchCharsetTest
-extends PHPUnit_Framework_TestCase
+extends \PHPUnit\Framework\TestCase
 {
     public function testQuery()
     {
@@ -36,11 +36,9 @@ extends PHPUnit_Framework_TestCase
         $this->assertFalse($s->query('iso-8859-1', true));
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testQueryWithoutBaseOb()
     {
+        $this->expectException(RuntimeException::class);
         $s = new Horde_Imap_Client_Data_SearchCharset();
 
         $s->query('UTF-8');
@@ -74,7 +72,7 @@ extends PHPUnit_Framework_TestCase
     {
         $s = new Horde_Imap_Client_Data_SearchCharset();
 
-        $mock = $this->getMock('SplObserver');
+        $mock = $this->createMock('SplObserver');
         $mock->expects($this->once())
             ->method('update')
             ->with($this->equalTo($s));
