@@ -24,13 +24,14 @@
  * @subpackage UnitTests
  */
 class Horde_Imap_Client_Socket_ClientSortTest
-extends PHPUnit_Framework_TestCase
+extends \PHPUnit\Framework\TestCase
 {
     public $socket_ob;
     public $sort_ob;
 
-    public function setUp()
+    public function setUp(): void
     {
+        parent::setUp();
         require_once __DIR__ . '/../Stub/ClientSort.php';
         require_once __DIR__ . '/../Stub/Socket.php';
 
