@@ -23,10 +23,11 @@
  * @package    Imap_Client
  * @subpackage UnitTests
  */
-class Horde_Imap_Client_Utf7ConvertTest extends PHPUnit_Framework_TestCase
+class Horde_Imap_Client_Utf7ConvertTest extends \PHPUnit\Framework\TestCase
 {
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
+        parent::setUpBeforeClass();
         require_once __DIR__ . '/Stub/Utf7imap.php';
     }
 
